Move ApiResponse & Twitter Status tests
